Add extra options specific to a solver
'--extra <filepath>' option is introduced.
This gives the path to a json file containing a dictionary.
This dictionary sets values for options that are specific to a
particular solver.

For instance, one extra option is introduced for defoCP and SRLS.
This argument controls the maximum number of intermediate segments
that these solvers can use.

This fixes Issue #7